HW2/clustering/requirements.txt
django
djangorestframework
markdown
django-filter
kubernetes